SDK: extract secret masking helper (oh-tab-g9ys)
